// src/entities/openapi/model/consentRequestDetailViewDto.ts
/**
 * agridata.ch API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ConsentRequestDetailViewDtoDataRequest } from './consentRequestDetailViewDtoDataRequest';
import { BurDto } from './burDto';
import { ConsentRequestStateEnum } from './consentRequestStateEnum';


/**
 * Data transfer object representing a consent request
 */
export interface ConsentRequestDetailViewDto { 
    id: string;
    /**
     * UID of the data producer
     */
    dataProducerUid?: string;
    /**
     * Current state of the consent request
     */
    stateCode?: ConsentRequestStateEnum;
    /**
     * Date and time when the state was changed last
     */
    lastStateChangeDate?: string;
    requestDate?: string;
    /**
     * BURs that belong to the consent request
     */
    burs?: Array<BurDto>;
    dataRequest?: ConsentRequestDetailViewDtoDataRequest;
}
export namespace ConsentRequestDetailViewDto {
}